requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
joblib
scikit-learn
geopy
pandas
numpy
